destoryコントローラ修正

+++ app/controllers/items_controller.rb
@@ -39,8 +39,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id != item.user_id
+      redirect_to root_path
+    else
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
